requirements.txt
crccheck==1.3.0
docopt==0.6.2
donkeycar==2.5.8
fastapi==0.95.0
filterpy==1.4.5
matplotlib==3.6.1
numpy==1.23.4
open3d==0.16.1
opencv_python==4.6.0.66
pandas==1.5.1
Pillow==9.4.0
pyrealsense2==2.53.1.4623
pyserial==3.5
uvicorn==0.21.1
